Build failures fixed for MC 1.12.2, just warnings left
