+++ README.md
@@ -1,0 +1,20 @@
+# Regrowable Leaves
+Minecraft Mod that slowly regrows leaves on trees.
+
+See https://minecraft.curseforge.com/projects/regrowable-leaves
+
+Leaves will slowly regrow over time when they are broken by a player.
+
+Leaves will only regrow if they are lit up "just like a tree growing from a sapling."
+
+They will not regrow if a tree has been cut down and leaves decay normally.
+
+*Now Supports Biome's O Plenty and Vanilla Leaves!*
+
+They will not regrow when the tree is cut down only if a player breaks the actual leaf blocks. Leaf decay still works as normal and regrown leaves will always check to make sure they can be sustained "Attached to a tree" or else the new leaves will decay.
+
+This works by placing an invisible block in the place of the original leaf block. If any other block is placed over this the leaf will not regrow.
+
+There is Config options to change the regrowth rate and light requirement.
+
+
