+++ build.gradle
@@ -1,46 +1,73 @@
-// For those who want the bleeding edge
+// build.gradle for mod
+// Based on build.gradle from Botania by vazkii - https://botaniamod.net/
+
+// START do not edit
 buildscript {
     repositories {
         jcenter()
-        maven {
-            name = "forge"
-            url = "http://files.minecraftforge.net/maven"
-        }
+        maven { url = "http://files.minecraftforge.net/maven" }
     }
     dependencies {
-        classpath 'net.minecraftforge.gradle:ForgeGradle:2.2-SNAPSHOT'
+        classpath 'net.minecraftforge.gradle:ForgeGradle:2.3-SNAPSHOT'
     }
 }
-apply plugin: 'net.minecraftforge.gradle.forge'
 
-/*
-// for people who want stable - not yet functional for MC 1.8.8 - we require the forgegradle 2.1 snapshot
 plugins {
-    id "net.minecraftforge.gradle.forge" version "2.0.2"
+  id "com.matthewprenger.cursegradle" version "1.0.10"
 }
-*/
-version = "1.10.2-3.1.2"
-group = "lilrichy.regrowableleaves" // http://maven.apache.org/guides/mini/guide-naming-conventions.html
-archivesBaseName = "regrowableleaves"
+
+apply plugin: 'net.minecraftforge.gradle.forge'
+apply plugin: 'com.matthewprenger.cursegradle'
+apply plugin: 'maven'
+// END do not edit
+//Only edit below this line, the above code adds and enables the necessary things for Forge to be setup.
+
+sourceCompatibility = targetCompatibility = '1.8' // Need this here so eclipse task generates correctly.
+compileJava {
+    sourceCompatibility = targetCompatibility = '1.8'
+}
+
+import groovy.json.JsonSlurper
+import groovy.json.JsonOutput
+
+ext.configFile = file('build.properties')
+
+ext.config = parseConfig(configFile)
+ext.priv = parseConfig(file('private.properties'))
+
 
 minecraft {
-    version = "1.10.2-12.18.1.2046"
-    runDir = "run"
+    version = "${config.mc_version}-${config.forge_version}"
+    runDir = "eclipse"
 
-    // the mappings can be changed at any time, and must be in the following format.
-    // snapshot_YYYYMMDD   snapshot are built nightly.
-    // stable_#            stables are built at the discretion of the MCP team.
-    // Use non-default mappings at your own risk. they may not allways work.
-    // simply re-run your setup task after changing the mappings to update your workspace.
-    mappings = "snapshot_20160518"
-    makeObfSourceJar = false // an Srg named sources jar is made by default. uncomment this to disable.
+    mappings = config.mc_mappings
+    // makeObfSourceJar = false // an Srg named sources jar is made by default. uncomment this to disable.
+
+    // This does the token replacement so names, version numbers, etc only need to be in one authoritative file
+    // TODO replace with a token such as @VERSION@
+    replace 'GRADLE:MOD_ID', config.mod_id
+    replace 'GRADLE:MOD_NAME', config.mod_name
+    replace 'GRADLE:BUILD', config.build_number
+    replace 'GRADLE:VERSION', config.version
+    replace 'GRADLE:INTERNALVERSIONMAJOR', config.internal_version_major
+    replace 'GRADLE:DEPENDENCIES', config.dependencies
+    replace 'GRADLE:UPDATEJSON', config.updatejson
+
+    replaceIn 'LibMisc.java'
+}
+
+version = config.version
+group = config.group // http://maven.apache.org/guides/mini/guide-naming-conventions.html
+archivesBaseName = config.mod_id
+
+sourceSets {
+    main {
+        java { srcDirs = ["$projectDir/src/java"] }
+        resources { srcDirs = ["$projectDir/src/resources"] }
+    }
 }
 
 dependencies {
-
-
-    compile files("G:/Users/Rich/Desktop/Regrowable Leaves/1.10.2/run/mods/BiomesOPlenty-1.9.4-4.1.0.2049-universal.jar")
-
     // you may put jars on which you depend on in ./libs
     // or you may define them like so..
     //compile "some.group:artifact:version:classifier"
@@ -50,14 +77,6 @@ dependencies {
     //compile 'com.mod-buildcraft:buildcraft:6.0.8:dev'  // adds buildcraft to the dev env
     //compile 'com.googlecode.efficient-java-matrix-library:ejml:0.24' // adds ejml to the dev env
 
-    // the 'provided' configuration is for optional dependencies that exist at compile-time but might not at runtime.
-    //provided 'com.mod-buildcraft:buildcraft:6.0.8:dev'
-
-    // the deobf configurations:  'deobfCompile' and 'deobfProvided' are the same as the normal compile and provided,
-    // except that these dependencies get remapped to your current MCP mappings
-    //deobfCompile 'com.mod-buildcraft:buildcraft:6.0.8:dev'
-    //deobfProvided 'com.mod-buildcraft:buildcraft:6.0.8:dev'
-
     // for more info...
     // http://www.gradle.org/docs/current/userguide/artifact_dependencies_tutorial.html
     // http://www.gradle.org/docs/current/userguide/dependency_management.html
@@ -65,21 +84,57 @@ dependencies {
 }
 
 processResources
-        {
-            // this will ensure that this task is redone when the versions change.
-            inputs.property "version", project.version
-            inputs.property "mcversion", project.minecraft.version
+{
+    // this will ensure that this task is redone when the versions change.
+    inputs.property "version", project.version
+    inputs.property "mcversion", project.minecraft.version
 
-            // replace stuff in mcmod.info, nothing else
-            from(sourceSets.main.resources.srcDirs) {
-                include 'mcmod.info'
+    // replace stuff in mcmod.info, nothing else
+    from(sourceSets.main.resources.srcDirs) {
+        include 'mcmod.info'
 
-                // replace version and mcversion
-                expand 'version': project.version, 'mcversion': project.minecraft.version
-            }
+        // replace version and mcversion
+        expand 'version':project.version, 'mcversion':project.minecraft.version
+    }
 
-            // copy everything else, thats not the mcmod.info
-            from(sourceSets.main.resources.srcDirs) {
-                exclude 'mcmod.info'
-            }
+    // copy everything else, thats not the mcmod.info
+    from(sourceSets.main.resources.srcDirs) {
+        exclude 'mcmod.info'
+    }
+}
+
+jar {
+    //rename the default output, for some better sanity with scripts
+    archiveName = "${baseName}-${minecraft.version}-${version}.${extension}"
+
+    exclude "**/*.bat"
+    exclude "**/*.psd"
+    exclude "**/*.exe"
+}
+
+/// Local functions
+
+def parseConfig(File config) {
+    config.withReader {
+        def prop = new Properties()
+        prop.load(it)
+        return (new ConfigSlurper().parse(prop))
+    }
+}
+
+/// CurseGradle
+
+curseforge {
+    apiKey = priv.cfkey
+    project {
+        id = config.curseforge_id
+        changelog = { ->
+            def updateJsonFile = new File("${projectDir}/${config.changelog}")
+            def updateJson = new JsonSlurper().parse(updateJsonFile) as Map
+            // this should error if changelog is missing change for this version - add one!
+            def changeinfo = new String(updateJson."${minecraft.version}"."${version}")
+            return changeinfo
         }
+        releaseType = "release"
+    }
+}
