+++ build.properties
@@ -1,0 +1,15 @@
+mod_name="Regrowable Leaves"
+forge_version=14.23.3.2698
+dir_repo=./
+mod_id=regrowableleaves
+mc_mappings=snapshot_20171217
+version=4.0.0
+dependencies=
+dir_output=../Build Output/RegrowableLeaves/
+internal_version_major=1
+curseforge_id=257612
+updatejson=https\://raw.githubusercontent.com/lilrichy/RegrowableLeaves/master/update.json
+group=lilrichy.regrowableleaves
+changelog=update.json
+mc_version=1.12.2
+build_number=356
