+++ private.properties.example
@@ -1,0 +1,2 @@
+cfkey=
+dir_forgecraft=../Build Output/RemovableLeaves/
