+++ src/java/com/blogspot/richardreigens/regrowableleaves/LogHelper.java
