+++ src/java/com/blogspot/richardreigens/regrowableleaves/blocks/ModBlocks.java
@@ -2,7 +2,10 @@ package com.blogspot.richardreigens.regrowableleaves.blocks;
 
 import com.blogspot.richardreigens.regrowableleaves.blocks.minecraftLevesBlock.BlockLeafAir;
 import com.blogspot.richardreigens.regrowableleaves.regrowableleaves;
-import net.minecraftforge.fml.common.registry.GameRegistry;
+
+import net.minecraft.block.Block;
+import net.minecraftforge.event.RegistryEvent;
+import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
 
 /**
  * Created by LiLRichy on 7/14/2016.
@@ -11,12 +14,8 @@ public class ModBlocks {
 
     public static final BlockLeafAir blockLeafAir = new BlockLeafAir();
 
-    public static void register() {
-        GameRegistry.register(blockLeafAir);
-
-        //We should only load these blocks if BOP is installed:
-        if (regrowableleaves.isBOPInstalled) {
-            ModBOPBlocks.register();
-        }
+    @SubscribeEvent
+    public void registerBlocks(RegistryEvent.Register<Block> event) {
+        event.getRegistry().registerAll(blockLeafAir);
     }
 }
