+++ src/java/com/blogspot/richardreigens/regrowableleaves/regrowableleaves.java
@@ -20,7 +20,6 @@ public class regrowableleaves {
     @Mod.EventHandler
     public void preInit(FMLPreInitializationEvent event) {
         ConfigurationHandler.init(event.getSuggestedConfigurationFile());
-        ModBlocks.register();
     }
 
     @EventHandler
