+++ src/resources/textlogo.png
@@ -1,0 +1,1 @@
+<html><body>You are being <a href="https://raw.githubusercontent.com/Stormwind99/RegrowableLeaves/master/src/main/resources/textLogo.png">redirected</a>.</body></html>
